+++ embark/workers/models.py
@@ -69,6 +69,7 @@ class Worker(models.Model):
     reachable = models.BooleanField(default=False)
     status = models.CharField(max_length=1, choices=ConfigStatus, default=ConfigStatus.UNCONFIGURED)
     analysis_id = models.UUIDField(blank=True, null=True, help_text="ID of the analysis currently running on this worker")
+    last_reached = models.DateTimeField(auto_now_add=True)
 
     dependency_version = models.OneToOneField(
         WorkerDependencyVersion,
@@ -91,15 +92,16 @@ class Worker(models.Model):
                     except ValueError as value_error:
                         raise ValidationError({"configuration": f"Invalid IP range: {value_error}"}) from value_error
 
-    def ssh_connect(self):
+    def ssh_connect(self, timeout=30):
         """
         Tries to establish an ssh connection with each configuration and returns the first successful connection
+        :param timeout: max ssh connect timeout
         """
         ssh_client = new_autoadd_client()
 
         for configuration in self.configurations.all():
             try:
-                ssh_client.connect(self.ip_address, username=configuration.ssh_user, password=configuration.ssh_password)
+                ssh_client.connect(self.ip_address, username=configuration.ssh_user, password=configuration.ssh_password, timeout=timeout)
 
                 # save the ssh user and password so they can later be used in commands
                 ssh_client.ssh_user = configuration.ssh_user
@@ -107,6 +109,9 @@ class Worker(models.Model):
                 break
             except (paramiko.SSHException, socket.error):
                 continue
+
+        if ssh_client.get_transport() is None or not ssh_client.get_transport().is_active():
+            raise paramiko.SSHException("Failed to connect to worker with any configuration.")
 
         return ssh_client